+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- Ability to export build list to CSV file has been added.
+
 ### Changed
 - Changed labels for automatic ticket handling.
 - Cleaned up context menu code.